requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.3
djangorestframework==3.10.1
idna==2.8
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
py-stackexchange==2.2.7
python-decouple==3.1
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
StackAPI==0.1.12
urllib3==1.25.3
